+++ src/router/routes.js
@@ -33,7 +33,7 @@ const routes = [
     path: "/creator-subscribers",
     component: () => import("layouts/FullscreenLayout.vue"),
     children: [
-      { path: "", component: () => import("src/pages/CreatorSubscribers.vue") },
+      { path: "", component: () => import("src/pages/CreatorSubscribersPage.vue") },
     ],
   },
   {
